feat : lv5 Exception + lv6 password Encrypt
lv5의 예외처리와 lv6의 비밀번호 암호화를 구현했다.

+++ src/main/java/com/example/schedulewithjpa/domain/User/service/UserService.java
@@ -4,8 +4,11 @@ import com.example.schedulewithjpa.domain.User.Entity.User;
 import com.example.schedulewithjpa.domain.User.dto.SignUpResponseDto;
 import com.example.schedulewithjpa.domain.User.dto.UserResponseDto;
 import com.example.schedulewithjpa.domain.User.repository.UserRepository;
+import com.example.schedulewithjpa.global.Config.PasswordEncoder;
+import com.example.schedulewithjpa.global.exception.NotMatchPwException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -13,9 +16,13 @@ public class UserService {
 
     public final UserRepository userRepository;
 
-    public SignUpResponseDto signUP(String username, String password, String email) {
+    public final PasswordEncoder passwordEncoder;
 
-        User user = new User(username, password, email);
+    public SignUpResponseDto signUp(String username, String password, String email) {
+
+        String encodedPassword = passwordEncoder.encode(password);
+
+        User user = new User(username, encodedPassword, email);
 
         User savedUser = userRepository.save(user);
 
@@ -24,21 +31,31 @@ public class UserService {
 
     public UserResponseDto findById(Long id) {
 
-        User findUser = userRepository.findByIdOrElseThrow(id);
+        User user = userRepository.findByIdOrElseThrow(id);
 
-        return new UserResponseDto(findUser.getUsername(), findUser.getEmail());
+        return new UserResponseDto(user.getUsername(), user.getEmail());
     }
 
-    public void delete(Long id) {
+    @Transactional
+    public void delete(Long id, String password) {
 
         User user = userRepository.findByIdOrElseThrow(id);
+
+        if(passwordEncoder.matches(password, user.getPassword())){
+            throw new NotMatchPwException("비밀번호가 일치하지 않습니다.");
+        }
 
         userRepository.delete(user);
     }
 
     public User login(String email, String password) {
 
-        return userRepository.findUserByEmailAndPasswordOrElseThrow(email, password);
+        User user = userRepository.findUserByEmailOrElseThrow(email);
 
+        if(passwordEncoder.matches(password, user.getPassword())){
+            throw new NotMatchPwException("비밀번호가 일치하지 않습니다.");
+        }
+
+        return user;
     }
 }
